+++ tests/conftest.py
@@ -1,7 +1,10 @@
 import pytest
 
-from notebooker.constants import DEFAULT_MONGO_DB_NAME, DEFAULT_RESULT_COLLECTION_NAME
+from notebooker.constants import DEFAULT_DATABASE_NAME, DEFAULT_RESULT_COLLECTION_NAME, DEFAULT_SERIALIZER
+from notebooker.settings import WebappConfig
 from notebooker.utils import caching
+from notebooker.utils.filesystem import initialise_base_dirs, _cleanup_dirs
+from notebooker.web.app import create_app, setup_app
 
 
 @pytest.fixture
@@ -16,7 +19,7 @@ def mongo_host(mongo_server):
 
 @pytest.fixture
 def test_db_name():
-    return DEFAULT_MONGO_DB_NAME
+    return DEFAULT_DATABASE_NAME
 
 
 @pytest.fixture
@@ -25,20 +28,22 @@ def test_lib_name():
 
 
 @pytest.fixture
-def template_dir(workspace, monkeypatch):
-    monkeypatch.setenv("TEMPLATE_DIR", workspace.workspace)
+def template_dir(workspace):
     return workspace.workspace
 
 
 @pytest.fixture
-def output_dir(workspace, monkeypatch):
-    monkeypatch.setenv("OUTPUT_DIR", workspace.workspace)
+def output_dir(workspace):
     return workspace.workspace
 
 
 @pytest.fixture
-def cache_dir(workspace, monkeypatch):
-    monkeypatch.setenv("CACHE_DIR", workspace.workspace)
+def cache_dir(workspace):
+    return workspace.workspace
+
+
+@pytest.fixture
+def py_template_dir(workspace):
     return workspace.workspace
 
 
@@ -52,12 +57,34 @@ def clean_file_cache(monkeypatch, workspace):
     caching.cache = None
 
 
-#
-# @pytest.fixture(autouse=True)
-# def _unset_notebooker_environ(config):
-#     """Remove Notebooker values from os.environ after each test."""
-#     yield
-#     for attribute in dir(config):
-#         if attribute.startswith("_"):
-#             continue
-#         os.environ.pop(attribute, None)
+@pytest.fixture()
+def webapp_config(mongo_host, test_db_name, test_lib_name, template_dir, cache_dir, output_dir, workspace):
+    return WebappConfig(
+        CACHE_DIR=cache_dir,
+        OUTPUT_DIR=output_dir,
+        TEMPLATE_DIR=template_dir,
+        SERIALIZER_CLS=DEFAULT_SERIALIZER,
+        SERIALIZER_CONFIG={
+            "mongo_host": mongo_host,
+            "database_name": test_db_name,
+            "result_collection_name": test_lib_name,
+        },
+        PY_TEMPLATE_BASE_DIR=workspace.workspace,
+        PY_TEMPLATE_SUBDIR="templates",
+    )
+
+
+@pytest.fixture
+def flask_app(webapp_config):
+    flask_app = create_app()
+    flask_app = setup_app(flask_app, webapp_config)
+    return flask_app
+
+
+@pytest.fixture
+def setup_and_cleanup_notebooker_filesystem(webapp_config):
+    try:
+        initialise_base_dirs(webapp_config=webapp_config)
+        yield
+    finally:
+        _cleanup_dirs(webapp_config)