+++ tests/integration/test_e2e.py
@@ -1,18 +1,13 @@
 # End to end testing
 import datetime
-import os
 
 import freezegun
 import git
-import mock
 import pytest
 
-from notebooker.constants import JobStatus, DEFAULT_SERIALIZER
-from notebooker.web.app import create_app
+from notebooker.constants import JobStatus
 from notebooker.web.routes.run_report import _rerun_report, run_report
 from notebooker.web.utils import get_serializer
-
-from ..utils import setup_and_cleanup_notebooker_filesystem
 
 DUMMY_REPORT = """
 # ---
@@ -43,36 +38,13 @@ cumulative.plot()
 """
 
 
-def _setup_workspace(workspace):
+@pytest.fixture
+def setup_workspace(workspace):
     (workspace.workspace + "/templates").mkdir()
     git.Git(workspace.workspace).init()
     (workspace.workspace + "/templates/fake").mkdir()
     report_to_run = workspace.workspace + "/templates/fake/report.py"
     report_to_run.write_lines(DUMMY_REPORT.split("\n"))
-
-
-def _environ(mongo_host, workspace, db_name, lib_name):
-    return {
-        "MONGO_HOST": mongo_host,
-        "MONGO_USERNAME": None,
-        "MONGO_PASSWORD": None,
-        "DATABASE_NAME": db_name,
-        "PY_TEMPLATE_DIR": workspace.workspace,
-        "GIT_REPO_TEMPLATE_DIR": "templates",
-        "RESULT_COLLECTION_NAME": lib_name,
-    }
-
-
-@pytest.fixture
-def environ(monkeypatch, mongo_host, workspace, test_db_name, test_lib_name):
-    """Setup workspace and environment variables for tests in this file."""
-    _setup_workspace(workspace)
-    update = _environ(mongo_host, workspace, test_db_name, test_lib_name)
-    for k, v in update.items():
-        if v is None:
-            monkeypatch.delenv(k, raising=False)
-        else:
-            monkeypatch.setenv(k, v)
 
 
 def _check_report_output(job_id, serialiser, **kwargs):
@@ -90,11 +62,9 @@ def _check_report_output(job_id, serialiser, **kwargs):
         assert getattr(result, k) == v, "Report output for attribute {} was incorrect!".format(k)
 
 
-@setup_and_cleanup_notebooker_filesystem
 @freezegun.freeze_time(datetime.datetime(2018, 1, 12))
-def test_run_report(bson_library, environ):
-    flask_app = create_app()
-    with flask_app.app_context() as c:
+def test_run_report(bson_library, flask_app, setup_and_cleanup_notebooker_filesystem, setup_workspace):
+    with flask_app.app_context():
         serialiser = get_serializer()
         overrides = {"n_points": 5}
         report_name = "fake/report"
@@ -102,7 +72,7 @@ def test_run_report(bson_library, environ):
         mailto = "jon@fakeemail.com"
         job_id = run_report(report_name, report_title, mailto, overrides, generate_pdf_output=False, prepare_only=True)
         _check_report_output(
-            job_id, serialiser, overrides=overrides, report_name=report_name, report_title=report_title, mailto=mailto,
+            job_id, serialiser, overrides=overrides, report_name=report_name, report_title=report_title, mailto=mailto
         )
         assert job_id == serialiser.get_latest_job_id_for_name_and_params(report_name, overrides)
         assert job_id == serialiser.get_latest_job_id_for_name_and_params(report_name, None)
@@ -110,47 +80,9 @@ def test_run_report(bson_library, environ):
         assert job_id == serialiser.get_latest_successful_job_id_for_name_and_params(report_name, None)
 
 
-@setup_and_cleanup_notebooker_filesystem
 @freezegun.freeze_time(datetime.datetime(2018, 1, 12))
-def test_run_report_old(bson_library, mongo_host, workspace, test_db_name, test_lib_name):
-    _setup_workspace(workspace)
-    for k, v in _environ(mongo_host, workspace, test_db_name, test_lib_name).items():
-        os.environ[k] = v
-    try:
-        flask_app = create_app()
-        flask_app.config["SERIALIZER_CLS"] = DEFAULT_SERIALIZER
-        flask_app.config["SERIALIZER_CONFIG"] = {"mongo_host": mongo_host, "mongo_db_name": test_db_name}
-        with flask_app.app_context() as c:
-            serialiser = get_serializer()
-            overrides = {"n_points": 5}
-            report_name = "fake/report"
-            report_title = "my report title"
-            mailto = "jon@fakeemail.com"
-            job_id = run_report(
-                report_name, report_title, mailto, overrides, generate_pdf_output=False, prepare_only=True
-            )
-            _check_report_output(
-                job_id,
-                serialiser,
-                overrides=overrides,
-                report_name=report_name,
-                report_title=report_title,
-                mailto=mailto,
-            )
-            assert job_id == serialiser.get_latest_job_id_for_name_and_params(report_name, overrides)
-            assert job_id == serialiser.get_latest_job_id_for_name_and_params(report_name, None)
-            assert job_id == serialiser.get_latest_successful_job_id_for_name_and_params(report_name, overrides)
-            assert job_id == serialiser.get_latest_successful_job_id_for_name_and_params(report_name, None)
-    finally:
-        for k, __ in _environ(mongo_host, workspace, test_db_name, test_lib_name).items():
-            del os.environ[k]
-
-
-@setup_and_cleanup_notebooker_filesystem
-@freezegun.freeze_time(datetime.datetime(2018, 1, 12))
-def test_run_report_and_rerun(bson_library, environ):
-    flask_app = create_app()
-    with flask_app.app_context() as c:
+def test_run_report_and_rerun(bson_library, flask_app, setup_and_cleanup_notebooker_filesystem, setup_workspace):
+    with flask_app.app_context():
         serialiser = get_serializer()
         overrides = {"n_points": 5}
         report_name = "fake/report"