+++ tests/utils.py
@@ -1,22 +1,6 @@
-import decorator
-
 from notebooker.settings import WebappConfig
-from notebooker.utils.filesystem import initialise_base_dirs, _cleanup_dirs
-from notebooker.utils.templates import get_all_possible_templates
-
-
-def setup_and_cleanup_notebooker_filesystem(f):
-    def blast_it(func, *args, **kwargs):
-        output, template, cache = None, None, None
-        try:
-            output, template, cache = initialise_base_dirs()
-            result = func(*args, **kwargs)
-            return result
-        finally:
-            conf = WebappConfig(OUTPUT_DIR=output, TEMPLATE_DIR=template, CACHE_DIR=cache)
-            _cleanup_dirs(conf)
-
-    return decorator.decorator(blast_it, f)
+from notebooker.web.app import create_app, setup_app
+from notebooker.web.utils import get_all_possible_templates
 
 
 def _gen_all_templates(template_dict):
@@ -28,5 +12,9 @@ def _gen_all_templates(template_dict):
 
 
 def _all_templates():
-    templates = list(_gen_all_templates(get_all_possible_templates(warn_on_local=False)))
-    return templates
+    web_config = WebappConfig(PY_TEMPLATE_BASE_DIR="")
+    flask_app = create_app()
+    flask_app = setup_app(flask_app, web_config)
+    with flask_app.app_context():
+        templates = list(_gen_all_templates(get_all_possible_templates(warn_on_local=False)))
+        return templates