+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.1.0 (2020-11-30)
+------------------
+Breaking changes
+
+* In config, PY_TEMPLATE_DIR has been renamed to PY_TEMPLATE_BASE_DIR
+* In config, GIT_REPO_TEMPLATE_DIR has been renamed to PY_TEMPLATE_SUBDIR
+
 0.0.2 (2020-10-25)
 ------------------
 Bugfixes & cleanup